+++ SpeedyWeather/src/SpeedyWeather.jl
@@ -184,6 +184,7 @@ include("parameterizations/radiation/shortwave_transmissivity.jl")
 include("parameterizations/radiation/clouds.jl")
 include("parameterizations/radiation/longwave_radiation.jl")
 include("parameterizations/radiation/longwave_transmissivity.jl")
+include("parameterizations/radiation/radiation.jl")
 include("parameterizations/radiation/greenhouse_gases.jl")
 include("parameterizations/stochastic_physics.jl")